requirements.txt
python-dotenv
fastapi==0.112.2
uvicorn[standard]==0.30.6
sqlalchemy==2.0.32
moviepy==1.0.3
pydantic==2.8.2
ffmpeg-python==0.2.0
whisper-ctranslate2==0.4.8
llama-cpp-python==0.2.90
requests==2.32.3
